+++ HeroIMDBClient/CoreData/MovieStorage.swift
@@ -32,7 +32,7 @@ extension MovieStorage: MovieStorageProtocol {
         movieQueue.sync {
             let context = self.store.persistentContainer.viewContext
             let request = NSFetchRequest<NSFetchRequestResult>(entityName: "MovieDB")
-            request.predicate = NSPredicate(format: "id == %i", Int16(movie.id))
+            request.predicate = NSPredicate(format: "id == %i", Int64(movie.id))
             request.returnsObjectsAsFaults = true
             if let results = try? context.fetch(request) as? [MovieDB],
                let firstSaved = results.first {
@@ -54,7 +54,7 @@ extension MovieStorage: MovieStorageProtocol {
     public func retrieve(movie id: Int) -> Movie? {
         movieQueue.sync {
             let request = NSFetchRequest<NSFetchRequestResult>(entityName: "MovieDB")
-            request.predicate = NSPredicate(format: "id == %i", Int16(id))
+            request.predicate = NSPredicate(format: "id == %i", Int64(id))
             request.returnsObjectsAsFaults = true
             let context = self.store.persistentContainer.viewContext
             if let results = try? context.fetch(request) as? [MovieDB],
@@ -80,7 +80,7 @@ extension MovieStorage: MovieStorageProtocol {
     public func retrieveAllMovies(_ page: Int) -> [Movie] {
         movieQueue.sync {
             let request = NSFetchRequest<NSFetchRequestResult>(entityName: "MovieDB")
-            request.predicate = NSPredicate(format: "page == %i", Int16(page))
+            request.predicate = NSPredicate(format: "page == %i", Int64(page))
             request.returnsObjectsAsFaults = true
             let context = self.store.persistentContainer.viewContext
             if let results = try? context.fetch(request) as? [MovieDB] {
